Fix some warnings, remove broken/incomplete support for K5, remove support for non-format VDIFB
git-svn-id: https://svn.atnf.csiro.au/difx@9289 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ libraries/mark5access/trunk/mark5access/blanker_mark5.c
@@ -328,7 +328,7 @@ int blanker_codif(struct mark5_stream *ms)
 		return 0;
 	}
 
-	data = (unsigned long long *)ms->payload;
+	data = (uint64_t *)ms->payload;
 
 	nword = ms->databytes/8;
